Fix edit for prevent duplicate image deleting

+++ app/Http/Controllers/ProductController.php
@@ -189,12 +189,11 @@ class ProductController extends Controller
         $validated = $request->validate([
             'category_id' => 'nullable|exists:categories,id',
             'name' => 'required|string|max:255',
-            'code' => 'nullable|max:50',
-            // 'code' => [
-            //     'string',
-            //     'max:50',
-            //     Rule::unique('products')->whereNull('deleted_at'),
-            // ],
+            'code' => [
+                'string',
+                'max:50',
+                Rule::unique('products')->whereNull('deleted_at'),
+            ],
             'size_id' => 'nullable|exists:sizes,id',
             'color_id' => 'nullable|exists:colors,id',
             'cost_price' => 'nullable|numeric|min:0',
@@ -234,7 +233,6 @@ class ProductController extends Controller
 
             // Create the product
             $product = Product::create($validated);
-            $product->update(['code' => 'PROD-' . $product->id]);
 
             // Add stock transaction if stock quantity is provided
             $stockQuantity = $validated['stock_quantity'] ?? 0; // Default to 0 if not provided
@@ -305,7 +303,6 @@ class ProductController extends Controller
             }
 
             $product = Product::create($validated);
-            $product->update(['code' => 'PROD-' . $product->id]);
 
             // Add stock transaction if stock quantity is provided
             $stockQuantity = $validated['stock_quantity'] ?? 0; // Default to 0 if not provided
@@ -401,7 +398,7 @@ class ProductController extends Controller
         $validated = $request->validate([
             'category_id' => 'nullable|exists:categories,id',
             'name' => 'string|max:255',
-            // 'code' => 'nullable|string|max:50',
+            'code' => 'nullable|string|max:50',
             // 'code' => 'string|max:50|unique:products,code,' . $product->id . ',id,deleted_at,NULL',
             'size_id' => 'nullable|exists:sizes,id',
             'color_id' => 'nullable|exists:colors,id',
@@ -417,8 +414,11 @@ class ProductController extends Controller
 
         // Handle image update
         if ($request->hasFile('image')) {
+            $imageUsageCount = Product::where('image', $product->image)
+                ->where('id', '!=', $product->id)
+                ->count();
             // Delete the old image if it exists
-            if ($product->image && Storage::disk('public')->exists(str_replace('storage/', '', $product->image))) {
+            if ($imageUsageCount === 0 && $product->image && Storage::disk('public')->exists(str_replace('storage/', '', $product->image))) {
                 Storage::disk('public')->delete(str_replace('storage/', '', $product->image));
             }
